requirements.txt
numpy==1.26.4
pydantic==2.10.6
soundfile==0.13.1
pydub==0.25.1
omegaconf==2.3.0
transformers==4.57.1
scipy==1.15.3
accelerate==1.11.0
torchaudio==2.2.2
torch==2.2.2
google-genai==1.46.0